+++ PaymentGateway.Models/MultiplePurchaseCommand.cs
@@ -8,7 +8,7 @@ namespace PaymentGateway.Models
         public class CommandDetails
         {
             public int ProductId { get; set; }
-            public decimal Quantity { get; set; }
+            public double Quantity { get; set; }
             // public bool RemoveFromPurchaseIfNotInStock { get; set; }
         }
